requirements.txt
click==8.0.1
joblib==1.0.1
nltk==3.6.2
numpy==1.21.0
pyspellchecker==0.6.2
regex==2021.4.4
tqdm==4.61.1
